+++ lib/src/test/java/me/aartikov/lib/data_binding/CommandTest.kt
@@ -1,0 +1,84 @@
+package me.aartikov.lib.data_binding
+
+import kotlinx.coroutines.test.runBlockingTest
+import me.aartikov.lib.utils.CoroutinesTestRule
+import me.aartikov.lib.utils.TestLifecycleOwner
+import org.junit.Assert
+import org.junit.Rule
+import org.junit.Test
+
+class CommandTest {
+
+    @get:Rule
+    val coroutinesTestRule = CoroutinesTestRule()
+
+    @Test
+    fun `doesn't receive command when not started`() = runBlockingTest {
+        val lifecycleOwner = TestLifecycleOwner()
+        val propertyObserver = TestPropertyObserver(lifecycleOwner)
+        val command = command<Int>()
+        val values = mutableListOf<Int>()
+        with(propertyObserver) { command bind { values.add(it) } }
+
+        command.send(0)
+
+        Assert.assertEquals(values, emptyList<Int>())
+    }
+
+    @Test
+    fun `receives command after starting`() {
+        val lifecycleOwner = TestLifecycleOwner()
+        val propertyObserver = TestPropertyObserver(lifecycleOwner)
+        val command = command<Int>()
+        val values = mutableListOf<Int>()
+        with(propertyObserver) { command bind { values.add(it) } }
+
+        command.send(0)
+        lifecycleOwner.onStart()
+
+        Assert.assertEquals(values, listOf(0))
+    }
+
+    @Test
+    fun `doesn't receive command when stopped`() {
+        val lifecycleOwner = TestLifecycleOwner()
+        val propertyObserver = TestPropertyObserver(lifecycleOwner)
+        val command = command<Int>()
+        val values = mutableListOf<Int>()
+        with(propertyObserver) { command bind { values.add(it) } }
+
+        lifecycleOwner.onStop()
+        command.send(0)
+
+        Assert.assertEquals(values, emptyList<Int>())
+    }
+
+    @Test
+    fun `saves commands after stopping`() {
+        val lifecycleOwner = TestLifecycleOwner()
+        val propertyObserver = TestPropertyObserver(lifecycleOwner)
+        val command = command<Int>()
+        val values = mutableListOf<Int>()
+        with(propertyObserver) { command bind { values.add(it) } }
+
+        lifecycleOwner.onStop()
+        repeat(3) { command.send(0) }
+        lifecycleOwner.onStart()
+
+        Assert.assertEquals(values, listOf(0, 0, 0))
+    }
+
+    @Test
+    fun `saves commands before starting`() {
+        val lifecycleOwner = TestLifecycleOwner()
+        val propertyObserver = TestPropertyObserver(lifecycleOwner)
+        val command = command<Int>()
+        val values = mutableListOf<Int>()
+        with(propertyObserver) { command bind { values.add(it) } }
+
+        repeat(3) { command.send(0) }
+        lifecycleOwner.onStart()
+
+        Assert.assertEquals(values, listOf(0, 0, 0))
+    }
+}
