+++ lib/src/test/java/me/aartikov/lib/data_binding/TestPropertyObserver.kt
@@ -1,0 +1,7 @@
+package me.aartikov.lib.data_binding
+
+import androidx.lifecycle.LifecycleOwner
+
+class TestPropertyObserver(
+    override val propertyObserverLifecycleOwner: LifecycleOwner
+) : PropertyObserver
